requirements.txt
# Core requirements
Pillow>=10.0.0  # For image handling
requests>=2.31.0  # For API requests
python-dateutil>=2.8.2  # For date handling

# Development requirements
pytest>=7.4.0  # For testing
black>=23.7.0  # For code formatting
flake8>=6.1.0  # For linting
